better structure and cleanup
