+++ custom_components/cloudweatherproxy/sensor.py
@@ -5,19 +5,14 @@ from __future__ import annotations
 from dataclasses import fields
 from .aiocloudweather.utils import resolve_caster
 import logging
-import time
 
 from .aiocloudweather import CloudWeatherListener, Sensor, WeatherStation
 
-from homeassistant.components.sensor import SensorEntity
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.device_registry import DeviceInfo
-from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import CloudWeatherProxyConfigEntry
-from .const import DOMAIN, UNIT_DESCRIPTION_MAPPING
+from .entity import CloudWeatherEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -65,71 +60,3 @@ async def async_setup_entry(
     cloudweather.new_dataset_cb.append(_new_dataset)
     entry.async_on_unload(
         lambda: cloudweather.new_dataset_cb.remove(_new_dataset))
-
-
-class CloudWeatherBaseEntity(Entity):
-    """The Cloud Weather Proxy Entity."""
-
-    _attr_has_entity_name = True
-    _attr_should_poll = False
-
-    sensor: Sensor | None
-    station: WeatherStation
-
-    def __init__(self, sensor: Sensor, station: WeatherStation) -> None:
-        """Construct the entity."""
-        self.sensor = sensor
-        self.station = station
-
-        self._attr_unique_id = f"{station.station_id}-{sensor.name}"
-        self._attr_device_info = DeviceInfo(
-            name=f"Weatherstation {station.station_id}",
-            identifiers={(DOMAIN, station.station_id)},
-            sw_version=station.station_sw_version,
-            manufacturer=getattr(station.vendor, "value", str(station.vendor)),
-        )
-        self._attr_available = (station.update_time is not None) and (
-            (station.update_time + 5 * 60) > time.monotonic())
-
-
-class CloudWeatherEntity(CloudWeatherBaseEntity, SensorEntity):
-    """The Cloud Weather Proxy Sensor Entity."""
-
-    def __init__(
-        self,
-        sensor: Sensor,
-        station: WeatherStation,
-        name: str,
-    ) -> None:
-        """Initialize the sensor entity."""
-        super().__init__(sensor, station)
-        self._attr_name = name
-        self._attr_native_value = sensor.value if sensor else None
-        description = UNIT_DESCRIPTION_MAPPING.get(sensor.unit)
-        if description is None:
-            for key, val in UNIT_DESCRIPTION_MAPPING.items():
-                try:
-                    if str(key) == sensor.unit:
-                        description = val
-                        break
-                except Exception:
-                    continue
-        if description is not None:
-            self.entity_description = description
-
-    async def update_sensor(self, station: WeatherStation) -> None:
-        """Update the entity."""
-        self.station = station
-        old_name = self.sensor.name if self.sensor is not None else None
-        if old_name is None:
-            return
-        new_sensor = getattr(station, old_name, None)
-        self.sensor = new_sensor
-
-        self._attr_native_value = self.sensor.value if self.sensor else None
-        self._attr_available = (station.update_time is not None) and (
-            (station.update_time + 5 * 60) > time.monotonic())
-
-        _LOGGER.debug("Updating %s [%s] with update time %s",
-                      self.unique_id, self.sensor, self.station.update_time)
-        self.async_write_ha_state()
